Add Receipts to Approved Cert Test Flows (#205)
* Fix inaccurate http response schemas

The app will crash if these schemas don't match the
actual format that Payments API returns.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Specify amounts in String instead of Double

I was getting issues when sending doubles over the wire.
The docs specify `doubles` / `floats` but for whatever
reason Strings worked. So this change aligns all amounts
to be expressed as Strings

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Drop unnecessary `merchantId` param

Since the `merchantId` is a passed in the `Merchant-Account`
header for nearly all of the API requests, we now pass
`merchantId` as a constructor param when creating
the PosApiService. We remove it here because it's
unused/redundant

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Expose fetched `merchant` obj conveniently

We ideally do not need to write this if-statement everywhere that
we want to access the fetched merchant value (or `null` if the
merchant was not fetched).

We can solve this problem by either modifying the sealed
`MerchantDetailsState` interface/class or by adding an accessor
like the one added in this commit.

The convenient accessor appears good enough for now but we can
revist in the future as our needs change

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Define net-new POS specific data models

So it turns out that PaymentsAPI is sending different shaped
responses for POS vs Online-Only flows. And given that we're
using retrofit to generate the response data accessors, we'll
start getting runtime errors if we continue to share the data
models. This is because fields that may exist or may not be
null for one could be the opposite in the other.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Create / organize receipt layout service

The original `pos/receipts` service defined three
level of abstraction for expressing the layout of a Receipt:
1) `ReceiptLinePart` - smallest unit of text in a layout. Akin
   to  a cell in table
2) `ReceiptLayoutLine` - consists of multiple
   `ReceiptLinePart`s. Akin to a row in a table
3) `ReceiptLayout` - consists of multiple `ReceiptLinePart`s
   and represents an entire receipt. This is the top-most
   level of the abstraction and can be physically printed or
   displayed in an Android View

These three models made it possible to represent receipts in
a common way. What still needed to be done - and what this
commit does - is define a convenient way of mapping data from
the Payments API into these receipt building blocks.

This commit defines a class for each test scenario in
certification. The class accepts constructor params that are
friendly towards the data that Payments API returns. Each class
also exposes a method for expressing this data as a
ReceiptLayout which can be easily viewed or printed.

This commit attempts to reduce duplication by
acknowledging that all test scenario receipts share
 a common template.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Hook up receipts for approved tx flows

This commit notably does not add support for receipts
in tx flows that encounter an error / unhappy path.
This commit does add support for receipts for all tx
flows in the approved scenarios / happy path

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>