fix issue with amount calculation in cashback purchase
It looks like the ebtCashAmounnt and cashBackAmounts are coming in as strings, then we're trying to add them, which will concatenate the strings instead of adding the numbers together.

Now this will cast each item to a double, do the math to get the total, then cast _back_ to a string to send to the backend.