only pass transaction type to receipts since that's all we need from the request
We were sending the entire payment request to the receipt generation code, which was more than we needed since we only used it to determine the transaction type.

This modifies things to instead only pass in the transaction type, which makes the rest of the code easier since we don't always have a full payment request to pass around in the case of refunds or voids.