look up payments & payment methods for out-of-band refunds & voids
We were relying on having a tokenizedPaymentMethod or a paymentRequest to render the receipts for voids and refunds. This worked if you had just created the payment and the app state didn't get reset, but if it did, those fields would be null and the app would crash since it couldn't render the receipt.

This will now look up the payment and paymentMethods as needed to ensure the receipt screens have the data they need without relying on stale values in memory.