Modify interfaces to use ISO 4 PinBlocks
Part of the ISO 4 spec for PIN blocks is to use aesEncryptEcb twice
in the process of deriving the PIN block. So the method name
`encryptPinBlock` was no longer appropriate. Since
AndroidKeyStoreStoredKey already implemented the aesEncryptEcb
method, for simplicity we opted to promote its method signature
to be the standard interface signature for aesEncryptEcb