Distinguish dukptClient vs workingKey txCount
The point of persisting the KSN file is to enable the dukpt client
to pick it's internal state exactly where it left off after calling
generateWorkingKey(). The matter of ensuring we send the correct
txCount associated with the workingkey in requests to rosetta proxy
is a separate matter.

Previously the KeySerialNumber class, which is serving as the way for
ForageTerminalSDK application code to interact with the txCounter
to send along to Rosetta Proxy, only exposed a `txCount` value. This
was causing an issue because it attempted to serve as both the value
that the DukptService would read to "pick up where it left off" and
it was the same txCount value that we were sending up with requests
to Rosetta Proxy. It cannot do both at once since those are different
values.

To address this issue, this commit splits `KeySerialNumber.txCount`
into two separate values:
* `KeySerialNumber.dukptClientTxCount` - to be used for
   DukptClient -> KSN file storage -> DukptClient
* `KeySerialNumber.workingKeyTxCount` - t be used as the txCount
   value sent up with the encrypted PIN block for rosetta proxy