Make Sample App POS-specific
There are two places that have dependencies on VGS and BT:
1. `ForageTerminalSDK`'s dependence on `ForageSDK`, which
    necessarily depends on BT and VGS for ecom txs
2. `ForagePINEditText`'s because this single class is
    currently responsible for handling both the ecom flows
    and the pos flows.

This commit takes the second step of addressing #1 by
refactoring deleting all usage of `ForageSDK` in the
sample app. The result is that there is no usage of
`ForageSDK` in the sample app anymore and, after the
previous commit, there is no usage of `ForageSDK` in
the pos repo anymore. This means the `ForageSDK` can
be safely deleted fulfilling #1

Once we decouple these modules, we can safely delete the
`ForageSDK` as it is unnecessary in the pos repo

NOTE: we delete several tests here as a shortcut to
ship this to Toast faster. If we kept the tests, there
would be a lot of additional scope to refactor them
to make this change.