Create `ForageElement` hierarchy
There are two places that have dependencies on VGS and BT:
1. `ForageTerminalSDK`'s dependence on `ForageSDK`, which
    necessarily depends on BT and VGS for ecom txs
2. `ForagePINEditText`'s because this single class is
    currently responsible for handling both the ecom flows
    and the pos flows.

This commit takes the first step of addressing #2 by
creating a `ForageElement` class hierarchy that decouples
`ForagePINEditText`'s dependence on `ForageSDK`:
* abstract ForageElement
  * abstract ForagePinElement
    * ForagePINEditText (pos; rosetta only)
    * ForagePINEditText (ecom; rosetta, bt, and vgs)
  * abstract ForagePanElement
    * ForagePANEditText (pos; logger specific to pos)
    * ForagePANEditText (ecom; logger specific to ecom)

There will only ever be 1 instance of `ForagePINEditText`
living in a repo and that instance will be tailored to
the flows of that repo (i.e. either pos flows or ecom
flows).

NOTE: in order to make the `Forage*EditText`s elements
repo-specific, we needed to make VaultWrapper generic to
the input-element it backs so that VaultWrapper is no
longer coupled to any specific vault (e.g. bt and vgs)