Expose showKeyboard method on ForageElements
Showing the keyboard requires different steps for the PAN
(a Forage-internal element) vs the vendor-backed PIN (not
an element we have access to directly).

VGS fortunately exposes a method to show the keyboard for
it's input. However, BT appears to lack a mechanism to
reveal the keyboard; hence the todo comment.

Signed-off-by: Devin Morgan <devin@joinforage.com>