Drop all pos related code from public ecom repo
We want the private pos repo and the public ecom repo to have the
ability to evolve separately accordingly their unique needs while
still sharing common code. The first step in this journey is to
make the public ecom repo completely unaware of the existence of
the private pos repo. We do that by deleting all of the pos code

Signed-off-by: Devin Morgan <devin@joinforage.com>