Bring back vaultToForageResponse with interfaces
Previously `vaultToForageResponse` relied on generics to because
VGS and BT responses differ. However, generics in Kotlin do not
support the union ("|") types, which means that using generics
will require the parent `AbstractVaultSubmitter` class to be aware
of all classes that implement it.

This is a problem because `AbstractVaultSubmitter` houses the
common template for of submitting requests to vaults and ought
to live in `/core`. But `/core` cannot be aware of any vault
specific details if it is to be reusable by the pos repo.

So, this commit, replaces the usage of generics with a common
interface that all VaultSubmitter may _privately_ implement. The
ability for VaultSubmitters to _privately_ implement the details
by which they interpret vault response and
ForageApiResponse.Success or ForageApiResponse.Failures makes it
possible to move `AbstractVaultSubmitter` to `/core` and still
house the common logic for submitting requests to vaults

Signed-off-by: Devin Morgan <devin@joinforage.com>