Increase flexibility of ForageElement hierarchy (#248)
* Drop all pos related code from public ecom repo

We want the private pos repo and the public ecom repo to have the
ability to evolve separately accordingly their unique needs while
still sharing common code. The first step in this journey is to
make the public ecom repo completely unaware of the existence of
the private pos repo. We do that by deleting all of the pos code

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Organize all files into core and ecom packages

The second step in the public ecommerce repo that needs to take
place in order to make sharing code between the private pos
repo and this repo pain free; all while allowing for both repos
to evolve independently

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Increase flexibility of ForageElement hierarchy

This is the third step in the public ecommerce repo that needs to
take place in order to make sharing code between the private pos
repo and this repo pain free; all while allowing for both repos
to evolve independently.

This hierarchy will make it easy to keep ecom things in the
ecom-ForagePinEditText and pos things in the pos-ForagePINEditText

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Bring back vaultToForageResponse with interfaces

Previously `vaultToForageResponse` relied on generics to because
VGS and BT responses differ. However, generics in Kotlin do not
support the union ("|") types, which means that using generics
will require the parent `AbstractVaultSubmitter` class to be aware
of all classes that implement it.

This is a problem because `AbstractVaultSubmitter` houses the
common template for of submitting requests to vaults and ought
to live in `/core`. But `/core` cannot be aware of any vault
specific details if it is to be reusable by the pos repo.

So, this commit, replaces the usage of generics with a common
interface that all VaultSubmitter may _privately_ implement. The
ability for VaultSubmitters to _privately_ implement the details
by which they interpret vault response and
ForageApiResponse.Success or ForageApiResponse.Failures makes it
possible to move `AbstractVaultSubmitter` to `/core` and still
house the common logic for submitting requests to vaults

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>