Split AbstractForageElement into interfaces
With the setForageConfig functionality moved into its own
service, are now able to split up AbstractForageElement
into atomic interfaces and, ultimate, delete it. Instead
of an abstract class that requires all subclasses to
implement all the things, we now have:
* interface DynamicEnvElement
* interface ForageElement<T: ElementState>
* interface EditTextElement
The since the new PinPad element will only live in the POS
repo, it only uses Rosetta and will not care about knowing
the ForageConfig (i.e. it's not a DynamicEnvElement). Also,
it won't be based on EditText so it won't implement
EditTextElement either; it only needs to implement
ForageElement<T: PinElementState>, which is great because
that's the only functionality ForagePinPad actually cares
about.

Signed-off-by: Devin Morgan <devin@joinforage.com>