Amend VaultWrapper work with Ecom and POS
For all Ecom flows, the ForageElement will have a ForageConfig
bound to it because the user must call setForageConfig ahead
of time. This is not true for all of the POS Elements. While
the PAN element in the POS repo will continue to expose a
setForageConfig so that it can know whether or not it is in
prod (we whitelist testing cards in if it's not in prod), the
PIN in the POS repo will not care at all about setForageConfig
because it only ever works with the Rosetta vault.

Before this commit, VaultWrapper facilities would not work with
POS-specific facilities because VaultWrapper had assumed that
the EnvConfig could be picked up from the ForagePINEditText.
This commit adds an additional param for EnvConfig so that
`getVaultSubmitter` can work with flows with UI elements that
do not have an EnvConfig bound to them (i.e. POS PIN flows)

Signed-off-by: Devin Morgan <devin@joinforage.com>