Abandon ForageSDKInterface interface
ForageSDKInterface started out as a nice-to-have if not
duplicative of the ForageSDK signature. ForageSDKInterface
could have had a purpose to communicate common facilities
exposed by both ForageSDK (ecom) and ForageTerminalSDK (pos)
but it's clear now that signatures between these SDKs have
divergent requirements: ecom is EditText-only UI + Redundant
Vaults while pos is both EditText and KeyPad UI + no vault
redundancy.

Since the text-collection mechanism is a param
in all of the submit methods, forcing both ForageSDK and
ForageTerminalSDK to both implement the same signatures means
both of these SDKs must accept the same interface or class
hierarchy of text-collection mechanisms (e.g.
ForagePinElement or SecurePinCollector). This is not acceptable
because we need to afford these SDKs to the flexibility of
different requirements.

Thus, ForageSDKInterface cannot be used across both Ecom
and POS and, at this point, it's there's a meaningful
maintenance burden to support docstrings in both. So,
this commit opts to delete it.

Signed-off-by: Devin Morgan <devin@joinforage.com>