Introduce SecurePinCollector interface
All of the common steps for submitting a pin live in
AbstractVaultSubmitter. The only ui-related steps common
to submitting a PIN are:
1. checking if the pin is complete
2. including the raw pin text in the vault request
3. clearing the raw pin text after sending the request

This commit has the AbstractVaultSubmitter access these
functionalities via an interface instead of specifically
a ForagePinElement to make AbstractVaultSubmitter
flexible enough to work with both EditText and the
new pos PinPad element.

Signed-off-by: Devin Morgan <devin@joinforage.com>