Create PosPinElementStateManager
A requirement for the ForagePinPad is that it needs to expose the raw
pin text length so that SDK consumers can show the appropriate number of
*** characters based on the number of digits a user has entered so far
as their PIN. ForagePinPad itself does not display this feedback.

The existing PinElementState remains sufficient for Ecom usecases
because all ecom flows rely on ForagePinElements, which are backed by
EditTexts. EditText display the number of *** characters so there is no
need to also know the length of the Pin. We also cannot know the length
of the PIN for Ecom since they use VGS and BT and these vendors do not
expose input length.

Part of this commit moves the ok-for-ecom facilities into core,
something that will need to be mirrored in the Ecom repo