Clean up ForageTerminalSDK
There are several reasons why ForageTerminalSDK ought to be cleaned up
and with modest urgency:
1. The file is almost 1000 LOC which makes it unwieldy to navigate
2. ForageTerminalSDK is the end software so it will necessarily require
integration tests. It is easier and cheaper to gain confidence in our
code by moving business logic into standalone services that can be
unittested than to need to make a ton of integration tests just to
ensure if-statements are getting hit correctly

Much of the cleanup involved moving log-statements into POS-specific
Repositories / Services. This address the immediate problem of moving
the if-statements and custom log messages into services rather than the
ForageTerminalSDK. However, this solution is not ideal either because
we're still passing down props that require integration tests (e.g.
datadog, OKHttpClient, etc all need to be need be instantiated or
mocked).

I have plans for a better solution and this solution becomes very very
evident with the cleanup changes made to ForageTerminalSDK in this
commit