Make TerminalSDK generic via ForageVaultElement
We need the POS to support EditText's in addition to Keypad based
collection mechanisms. The point of creating the ForageVaultElement was
to avoid needing to duplicate each ForageTerminalSDK method where one
would service EditText-based inputs and the other would service PinPad
based inputs.

Things to NOTE:
* to reduce needing to specify POS pin state everywhere, we create the
ForagePosVaultElement typealias
* the ForagePinPad stores the pin-text differently from the EditText.
So, we refactor RosettaPinSubmitter to accept the pin's plaintext as
that is a common way to bind the plaintext so it can remain decoupled to
the backing instrument