Add ForagePinPad to Sample app
In the process we can also stop pass or set ForageConfig since the
ForagePinPad no longer needs it.

NOTE: this is a very very crude way to introduce the ForagePinPad. We're
just slapping it in there. This commit does not even add a UI to display
the ***s as a user types. You can still use the keypad though you just
won't see any visual feedback.