Correctly style ForagePinPad in Sample App
An Android view must be inflated in order for the attrs it's passed via
constructor param to have the correct mapping of styleable indices
pointing to resource IDs.

I was previously trying to render a ForagePinPad instance via
AndroidView in the ComposableForagePinPad and this was not interpreting
the <declare-styleable name="ForageKeypad"> correctly. The right was to
render the ForagePinPad is to render an instance of it in XML, which
I've done with forage_pin_pad_example.xml and use AndroidViewBinding. I
had to include the "androidx.compose.ui:ui-viewbinding:1.3.2" dependency
for this work.

NOTE: I opted to update the format's for of the text, height, and margin
attrs within ForageKeypad so that they reflect their intended types
instead of leaving them as references