[FX-1396] Get rid of `AbstractForageElement` in favor of interfaces (#254)
* Move ForageConfig init logic to Manager class

We need to replace (delete) AbstractForageElement with
smaller component interfaces in order to introduce the
PinPad UI component. We find ourselves needing the
ability to separately declare that an element can
implement any combination of facilities such as the
below without needing to be forced to implement ones
the Element does not need:
* facilities common to all Forage UI
* facilities common to only EditText-based UI
* facilities common to only UI that care about the env

This commit takes the first step of splitting up
AbstractForageElement into more atomic interfaces by first
migrating the setForageConfig related logic to it's own
service

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Split AbstractForageElement into interfaces

With the setForageConfig functionality moved into its own
service, are now able to split up AbstractForageElement
into atomic interfaces and, ultimate, delete it. Instead
of an abstract class that requires all subclasses to
implement all the things, we now have:
* interface DynamicEnvElement
* interface ForageElement<T: ElementState>
* interface EditTextElement
The since the new PinPad element will only live in the POS
repo, it only uses Rosetta and will not care about knowing
the ForageConfig (i.e. it's not a DynamicEnvElement). Also,
it won't be based on EditText so it won't implement
EditTextElement either; it only needs to implement
ForageElement<T: PinElementState>, which is great because
that's the only functionality ForagePinPad actually cares
about.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Amend VaultWrapper work with Ecom and POS

For all Ecom flows, the ForageElement will have a ForageConfig
bound to it because the user must call setForageConfig ahead
of time. This is not true for all of the POS Elements. While
the PAN element in the POS repo will continue to expose a
setForageConfig so that it can know whether or not it is in
prod (we whitelist testing cards in if it's not in prod), the
PIN in the POS repo will not care at all about setForageConfig
because it only ever works with the Rosetta vault.

Before this commit, VaultWrapper facilities would not work with
POS-specific facilities because VaultWrapper had assumed that
the EnvConfig could be picked up from the ForagePINEditText.
This commit adds an additional param for EnvConfig so that
`getVaultSubmitter` can work with flows with UI elements that
do not have an EnvConfig bound to them (i.e. POS PIN flows)

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Abandon ForageSDKInterface interface

ForageSDKInterface started out as a nice-to-have if not
duplicative of the ForageSDK signature. ForageSDKInterface
could have had a purpose to communicate common facilities
exposed by both ForageSDK (ecom) and ForageTerminalSDK (pos)
but it's clear now that signatures between these SDKs have
divergent requirements: ecom is EditText-only UI + Redundant
Vaults while pos is both EditText and KeyPad UI + no vault
redundancy.

Since the text-collection mechanism is a param
in all of the submit methods, forcing both ForageSDK and
ForageTerminalSDK to both implement the same signatures means
both of these SDKs must accept the same interface or class
hierarchy of text-collection mechanisms (e.g.
ForagePinElement or SecurePinCollector). This is not acceptable
because we need to afford these SDKs to the flexibility of
different requirements.

Thus, ForageSDKInterface cannot be used across both Ecom
and POS and, at this point, it's there's a meaningful
maintenance burden to support docstrings in both. So,
this commit opts to delete it.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Apply spotless

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Make element interfaces internal

These interfaces are, more than anything, a way for Forage developers
to organize methods for our public components. The compiler supports
these methods being public on the classes that implement these
interfaces, so it's not necessary that the interfaces themselves
actually be public.

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>