[FX-1383] Use `SecurePinCollector` to decouple PinSubmitters from `ForagePINEditText` (#255)
* Introduce SecurePinCollector interface

All of the common steps for submitting a pin live in
AbstractVaultSubmitter. The only ui-related steps common
to submitting a PIN are:
1. checking if the pin is complete
2. including the raw pin text in the vault request
3. clearing the raw pin text after sending the request

This commit has the AbstractVaultSubmitter access these
functionalities via an interface instead of specifically
a ForagePinElement to make AbstractVaultSubmitter
flexible enough to work with both EditText and the
new pos PinPad element.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Properly reference the clearText() of Submitter

This was causing an infinite loop before as `this` referred to the
object rather than the class

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>