Delete some unnecessary and unused code (#256)
* Delete some unnecessary code

Either this code is not needed at all or it should only
live in the pos repo. In either case, this code is not
needed in the ecom repo so we delete it

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Drop determineBackingVault from ForagePinElement

determineBackingVault is only relevant for classes that implement
DynamicEnvElement. Since the PIN in the pos repo will not by a
DynamicEnvElement, it no longer makes sense for determineBackingVault to
live in the common ForagePinElement. Instead it should just live in the
ecom-specific ForagePINEditText

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Apply Spotless

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>