[FX-1460] Insert ForageVaultElement into PIN class hierarchy (#267)
* Insert ForageVaultElement into PIN class hierarchy

ForageVaultElement is basically the union type ForagePinElement and
ForagePinPad (pos repo). To realize this union type, we need to ensure
that ForagePinElement subclasses ForageVaultElement.

The main complexity of this initiative came from ForageVaultElement
being a generic that requires knowing the shape of the state that will
be passed in on-change events, focus / blur, etc. We now have (or
rather, need) three different states:
1. PanEditTextState
2. PinEditTextState
3. PinPadState (pos repo only)

The problem was that PinElementStateManager was not flexible enough to
support a PinEditTextState and a PinPadState, and this meant we needed
to re-home the state tracking functionality into more atomic concepts
such as FocusState and PinInputState; teasing out these separate
concerns gave us the the flexibility we needed to support all three new
states outlined above

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Add missing docstrings

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>