Zero all padding for BT, VGS, and Rosetta elements (#274)
<!-- Update your title to prefix with your ticket number -->

## What
Ensure the base appearances of the VGS, BT, and Rosetta text elements are identical. The main problem was that VGS and BT were setting default patterns. This PR zeroes out the paddings of all of our vaults.

It's worth mentioning that padding has the identical visual effect as adjusting the width and height of the text element when the text is horizontally and vertically centered. For this reason, I believe zeroing out the padding is an acceptable solution.


<!-- Please include a summary of the change. List any dependencies that are required for this change. -->

## Why
Forage needs to route traffic to different vault providers transparently based on availability. We were facing a problem where swapping out the vault providers was _not_ transparent when `inputHeight` or `inputWidth` attributes were not supplied to `ForagePINEditText`, a valid SDK usage.

<!-- Describe the motivations behind this change if they are a subset of your ticket -->

## Test Plan
This was a purely appearance change, so I did not add any tests.

## Demo
NOTE: The demo images are from two other branches that modified `ForagePINEditText` to render VGS, BT, and Rosetta side-by-side. We don't want those demo-changes in `main` so that demo-code lives in separate branches that won't get merged in (i.e. #275 and #276)

### Before - No Styling Parity

> 📌 To view this demo for yourself, pull `devin/demo-with-NO-styling-parity`

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IfyQmT9sZxXL1kIgYI4y/8225aa50-685b-4ab2-816b-f652d17e73b6.png)

### After - With Styling Parity

> 📌 To view this demo for yourself, pull `devin/demo-with-styling-partity`

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IfyQmT9sZxXL1kIgYI4y/7b80f2ca-d2e8-473e-bbaa-c52e2113f300.png)


<!-- If applicable, please include a screenshot to this PR description -->
<!-- If applicable, please include a screen recording and post it in #feature-recordings -->

## How
Let's discuss what we think about this change. Once we're all aligned, this PR can be merged in.
NOTE: I will be closing out #275 and #276. They were only meant to serve as demos and not actually meant to be merged.
<!-- Describe the rollout plan if it includes multiple PRs/Repos or requires extra steps beyond reverting this PR -->