Reduce verbosity of Metrics facilities
We've been using several Metrics for some time and what's emerging is
that there is real need to decouple the Metric's construction with it's
"start"ing of the timer. We can make the calling code more succinct
by dropping the requirement that construction and start must happen
separately.

This commit also replaces some factory methods with simple constructor
calls. It also makes setEventOutcome a class method because it's logic
is the responsibility of CustomerPerceivedResponseMonitor and not of
ForageTerminalSDK