Delete inappropriate files
In our attempt to bring the changes from
`devin/all-the-things-for-forage-pin-pad` into `main`, resolving
changes on that branch with changes from `upstream/main` is revealing a
number of files that should no longer exist in the Pos `main` branch.
All of these files have either been moved or ought to be deleted.