Have ForageTerminalSDK.init to accept ksnDir param
For security, the ideal would be that Forage's SDK never has access to
the `context` object. However, all views are bound to a context in
Android so Forage will have access to the `context` object.

In lieu of this security ideal, we at least want to empower host
applications to specify the directory the KSN file should reside to
minimize the chances that the host app interferes with the KSN file