Move two methods up one level in class hierarchy
Both `showKeyboard()` and `getVaultSubmitter()` only interact with the
`.vault` and do not care about the concrete backing vault. This suggests
these methods would more appropriately live in `ForagePinElement`.

This will allow us to avoid needing to declare these methods in both the
ecom and the pos implementations of `ForagePINEditText`, which is a
small reusability win

Signed-off-by: Devin Morgan <devin@joinforage.com>