Organize ForageApiResponse.Failure logic (#290)
* Organize ForageApiResponse.Failure logic

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Apply spotless

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Update Rosetta test to support more granular error

This PR adds support for parsing errors from Rosetta, Payments API etc
with more granularity. In this test case, an error that was previously
not parsed (instead we just threw an exception) will no be parsed and
pleasantly returned to the SDK user

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Use @deprecated tag for `.errors`

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Fix some typos

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Clarify when Rosetta bad requests / responses comments

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Ensure API-VERSION header set to default

It's not appropriate to drop this value in this PR because this PR is
only responsible for re-organizing the ForageApiResponse.Failure logic.
We'll move this change to the next PR where we add sync endpoints
because that's when it's appropriate to do this

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Clarify error-parsing logic is for error-payloads

This will hopefully clarify that these classes only refer to the payload
so they are not expected to contain the status codes. Rather it's the
`ForageError` that will contain the status code

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Have ForageErrorDetails live in .polling

We faced a tradeoff of prioritizing an organized repo or happy forage
partners. Forage is now at a point where we have to prioritize happy
Forage partners

This PR moves the ForageErrorDetails back to it's, unfortunate, original
home so that we avoid introducing a breaking change while releasing sync
endpoints.

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>