Make Balance and Capture use sync endpoints (#288)
* Insert sync endpoints in BT and Rosetta

Note that we are not deleting any polling code here - we're just
inserting the sync endpoints into the flow. We are only deleting test
cases testing internal details of how pin-submission works as SQS
messages are simply not involved in the flow with sync-endpoints

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Add BT parsing tests

BT returns different shaped responses with API-VESION: 2024-01-08. To
ensure reliability, it's time to add unit test for these facillities.

* Apply spotless

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Catch and log BT response parsing exceptions

We have the ability now to differentiate between a failure to parse
what should be a ForageError and a failure to parse what should be a
successful response from BT.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Add comments for balance type back-and-forth

It took me a very long time to understand why we go back and forth
between ForageApiResponse.Success and EbtBalance so I'm adding this
comments to clarify for other developers

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Re-drop setting API-VERSION: default in AbstractVaultSubmitter

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Get rid of unnecessary when statements and logs

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Fix typo

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>