Mark setDoneListener deprecation as WARNING
It was accidentally set as ERROR and Toast was still using that. This
slipped below our radar because we did not use the Done button in our
Sample App. So, this commit now adds the Done button to our sample app
so that we'll have the same experience as Toast