Prune emvchip service's responsibilities
We decided that Payments API will have complete say of interpreting the
track 2 data and deciding whether a card is legacy vs fallback. So, we
should get rid of all Android client code that interprets the track 2
data. Also, let's drop all the EMV and Field 55 related things until we
cross that bridge.