Lazy init keyPair in RsaKeyManager
Creating the KeyPair takes almost 10s. We were running this operation
every time .init() was called regardless of whether we decided to exit
the KSN initialization process early or not because this was happening
inside the class constructor. Moving this logic into a lazy init var
allows us to instantiate RsaKeyManager instances without paying this
cost and only paying this cost in the case where a full KSN init is
necessary.

As part of this change, we also drop the deleteKeyPair operation since
it will get garbage collected