Modify the PinSubmission pattern to support Ecom+Pos
We recently changed POS so that partners will not need to create
PaymentMethod before performing balance checks or payment capture. This
means that obtaining the paymentMethodRef should no longer be part of
the common PinSubmission flow across both Pos and Ecom; instead it is
specific to Ecom.

This work adapts the PinSubmission pattern to support Ecom (net-new) in
a way that makes PinSubmission no longer aware of paymentMethods.

Signed-off-by: Devin Morgan <devin@joinforage.com>