Separate Ecom vs Pos error parsing strategies
As much as possible we want Ecom and Pos to be decoupled so they can
evolve according to their unique needs. We noticed that the Pos deals
with refund-related errors while Ecom does not motivating the splitting
up of these error response parsing. Since the HttpEngine is tightly
coupled to the error response parsing strategy, this meant creating two
seperate OkHttpEngines one for Pos and one for Ecom