Address Danny's feedback
Balance Checks and Deferred capture we don't want to use idempotency
protection since money is not moving, so these should use a UUID. For
sync capture, we want idempotency protection since money would move
immediately. Technically, passing the payment-ref is not ideal since it
lacks the resolution of "individual attempt sessions". BUT, this puts us
at parity with the existing SDK and avoids changing behavior

Signed-off-by: Devin Morgan <devin@joinforage.com>